Reflow as a left nav and map taking up rest of space
This moves things around, replacing the app with a left nave containing our title, description, and toggles, and letting the map take up the rest of the space.

This accomodates mobile pretty well, so was able to remove our special mobile javascript code. Was also able to remove a good amount of CSS that was managing embedding the map in our nav, and resizing the map.

There's a need to be able to toggle the nav's visibility on smaller screens which I'm leaving as a separate issue.